pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "open-problem-atlas"
version = "0.1.0"
description = "A living, machine-readable atlas of open problems for researchers, AI agents, and theorem provers."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pre-commit>=3.0",
    "ruff>=0.1.0",
]
ingestion = [
    "httpx>=0.25",
    "beautifulsoup4>=4.12",
    "lxml>=4.9",
]
site = [
    "jinja2>=3.1",
]
snapshot = [
    "pyarrow>=14.0",
]
runner = [
    "pyyaml>=6.0",
    "jsonschema>=4.20",
]
runner-smt = [
    "z3-solver>=4.12",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
